ai-engine/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
pydantic
python-dotenv
requests
python-multipart
pypdf
google-generativeai
google-genai
